Behat - add a new configuration key in behat.yml file : user_id_to_not_delete So you can use an existing dotclear installation and not delete superadmin user
Use cookie session name provide in dotclear configuration file

+++ features/sql/db_cleanup.sql
@@ -1,1 +1,1 @@
-DELETE FROM dc_user;
+DELETE FROM dc_user WHERE user_id != '__USER_ID__';
